[refactor] #35 auth 패키지 리팩토링
