+++ .github/workflows/deploy.yml
@@ -20,6 +20,8 @@ jobs:
           echo "${{ secrets.APPLICATION_JWT_YML }}" > ./application-jwt.yml
           echo '${{ secrets.APPLICATION_EMAIL_YML }}' > ./application-email.yml
           echo '${{ secrets.APPLICATION_REDIS_YML }}' > ./appication-redis.yml
+          echo '${{ secrets.APPLICATION_OAUTH_YML }}' > ./appication-oauth.yml
+
         shell: bash
 
       - name: Set up Docker Buildx
