+++ src/main/java/com/example/dgu/returnwork/domain/auth/dto/response/GoogleLoginResponseDto.java
@@ -8,18 +8,21 @@ public record GoogleLoginResponseDto(
 
         String accessToken,
 
-        String refreshToken
+        String refreshToken,
+
+        String tempToken
 ){
 
     public static GoogleLoginResponseDto login(String accessToken, String refreshToken) {
-        return new GoogleLoginResponseDto("SUCCESS", "로그인 성공", accessToken, refreshToken);
+        return new GoogleLoginResponseDto("SUCCESS", "로그인 성공", accessToken, refreshToken, null);
     }
 
-    public static GoogleLoginResponseDto signUpNeeded() {
+    public static GoogleLoginResponseDto signUpNeeded(String tempToken) {
         return new GoogleLoginResponseDto(
                 "SIGNUP_NEEDED",
                 "추가 정보를 입력해주세요",
                 null,
-                null);
+                null,
+                tempToken);
     }
 }
