+++ src/main/java/com/example/dgu/returnwork/domain/user/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.example.dgu.returnwork.domain.user.controller;
 
-import com.example.dgu.returnwork.domain.user.dto.request.SignUpRequestDto;
 import com.example.dgu.returnwork.domain.user.dto.request.VerifyEmailRequestDto;
 import com.example.dgu.returnwork.domain.user.service.UserCommandService;
 import com.example.dgu.returnwork.domain.user.service.UserQueryService;
@@ -9,17 +8,13 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/auth")
+@RequestMapping("/api/users")
 public class UserController implements UserApi {
 
     private final UserCommandService userCommandService;
     private final UserQueryService userQueryService;
 
-    @Override
-    @PostMapping("/signup")
-    public void signUp(SignUpRequestDto request) {
-        userCommandService.signUp(request);
-    }
+
 
     @Override
     @GetMapping("/duplicate")
@@ -38,5 +33,4 @@ public class UserController implements UserApi {
     public void verifyEmail(VerifyEmailRequestDto request) {
         userQueryService.verifyEmail(request);
     }
-
 }
