+++ src/main/java/com/example/dgu/returnwork/global/properties/JwtProperties.java
@@ -23,6 +23,6 @@ public class JwtProperties {
     }
 
 
-    public record Expiration(Long access, Long refresh) {
+    public record Expiration(Long access, Long refresh, Long temp) {
     }
 }
